chore: fix model imports
